requirements.txt
click==8.1.3
Flask==2.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycrypto==2.6.1
pycryptodome==3.15.0
Werkzeug==2.1.2
